requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Smorest
marshmallow
python-dotenv
sgp4
numpy
requests
# Add other dependencies as needed